modifications du seeder de la bdd

+++ back/database/migrations/2025_10_28_134357_badge_migration.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('badges', function (Blueprint $table) {
             $table->id();
-            $table->integer('numero')->unique();
+            $table->unsignedBigInteger('numero')->unique();
             $table->timestamps();
         });
     }
